initial bigfile read support
Needs more tests, but it seems to work...

I seem to have discovered a bug in mogtool which causes it to
generate incorrect MD5 checksums when the --gzip flag is used
(and --gzip actually just does zlib deflate, not something that
gzip(1) can actually decrypt).

So right now MD5 checksums are only verified on non-zlib-deflated
files.