MogileFS::HTTPFile: support TCP_CORK on Linux
Since we use TCP_NODELAY with all sockets, the bigfile code path
here is less efficient because it sends the headers separately.
So use TCP_CORK for this part of the code to avoid sending
partial frames.

Additionally rewrite this to use send(2) which is more suitable
for sockets.

Since we didn't have any previously, add a test for uploading
large files.