Use syswrite_full when PUT-ing smaller files
Socket#send will return with short writes and IO.write just
sucks, so expose our existing syswrite_full routine to the
public.

While we're at it, add a unit test to try to ensure that we're
correctly doing syswrite_full here.  System-specific behavior
dealing with buffer sizes is not easy to test...