backend: don't blindly add "Error" to exception class names
We were getting classes with "SizeVerifyErrorError", which
just looks stupid.  This was because MogileFS puts
the "_error" suffix into error messages it sends us and
we were blindly appending Error to the end.  Avoid doing
that, even if it's somewhat inconsistent since "ErrorError"
looks really stupid IMNSHO...

Note: this is a minor API change